+++ test/index.test.js
@@ -1,0 +1,8 @@
+import { expect } from 'chai';
+import * as Rasti from '../src/index.js';
+
+describe('Rasti', () => {
+    it('must exists', () => {
+        expect(Rasti).to.exist;
+    });
+});
